Correção de bugs

+++ backend/src/main/java/br/com/positivo/brilhamais/models/FaixaPontuacao.java
@@ -28,5 +28,5 @@ public class FaixaPontuacao {
     private BigDecimal valorMaximo;
 
     @Column(name = "pontos_obtidos")
-    private Integer pontosObtidos;
+    private BigDecimal pontosObtidos;
 }
